+++ NEWS.md
@@ -1,3 +1,7 @@
+## PxWebApiData	0.6.0
+
+* Changes to meet to meet CRAN policy:
+  - *Packages which use Internet resources should fail gracefully with an informative message if the resource is not available or has changed (and not give a check warning nor error).*
 
 ## PxWebApiData	0.5.0